requirements.txt
numpy
torch 
transformers
accelerate
tqdm
hydra-core
sentencepiece
mauve-text